services/llm-service/requirements.txt
fastapi
psycopg2-binary
google-generativeai
